requirements.txt
aiohttp==3.9.3
aiosignal==1.3.2
asyncio==3.4.3
attrs==24.3.0
aws-lambda-powertools==1.25.5
aws-xray-sdk==2.14.0
boto3==1.36.2
botocore==1.36.2
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
clr_loader==0.2.7.post0
fastjsonschema==2.21.1
frozenlist==1.5.0
fyers_apiv3==3.1.4
idna==3.10
jmespath==0.10.0
multidict==6.1.0
numpy==2.0.1
pandas==2.2.3
propcache==0.2.1
pycparser==2.22
python-dateutil==2.9.0.post0
pythonnet==3.0.5
pytz==2024.2
PyYAML==6.0.2
requests==2.31.0
s3transfer==0.11.1
six==1.17.0
stock-indicators==1.3.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
websocket-client==1.6.1
wrapt==1.17.2
yarl==1.18.3
